add user validation in use cases and add some tests

+++ src/main/kotlin/domain/usecase/task/DeleteUserTaskUseCase.kt
@@ -1,9 +1,16 @@
-package danis.galimullin.domain.usecase.task
+package domain.usecase.task
 
+import danis.galimullin.domain.exception.ForbiddenException
 import domain.repository.TaskRepository
+import io.ktor.server.plugins.NotFoundException
 
 class DeleteUserTaskUseCase(private val repository: TaskRepository) {
-    suspend operator fun invoke(taskId: Long) {
+    suspend operator fun invoke(taskId: Long, userId: Long) {
+        val task = repository.getById(taskId) ?: throw NotFoundException()
+
+        if(userId != task.userId) {
+            throw ForbiddenException()
+        }
         repository.delete(taskId)
     }
 }
